# backend/api-docs.yaml
openapi: 3.1.0
info:
  title: SSD-CVE API
  description: "Self-Sovereign Digital Credential Verification Engine. Register, issue,\
    \ hold, verify, and revoke verifiable credentials."
  version: 1.0.0
servers:
- url: http://localhost:6969
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/verifier/verify:
    post:
      tags:
      - verifier-controller
      operationId: verify
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                credentialFile:
                  type: string
                  format: binary
              required:
              - credentialFile
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
  /api/verifier/verify/batch:
    post:
      tags:
      - verifier-controller
      operationId: verifyBatch
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchVerificationResponse"
  /api/verifier/verify/batch/csv:
    post:
      tags:
      - verifier-controller
      operationId: exportCsv
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchVerificationResponse"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
...
  /api/verifier/verify/{credentialId}:
    get:
      tags:
      - verifier-controller
      operationId: verifyById
      parameters:
      - name: credentialId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
...
  /api/verifier/anchors/recent:
    get:
      tags:
      - verifier-controller
      operationId: recentAnchors
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentAnchorResponse"
  /api/verifier/anchor/{credentialNumber}:
    get:
      tags:
      - verifier-controller
      operationId: anchor
      parameters:
      - name: credentialNumber
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnchorLookupResponse"
...
components:
  schemas:
...
    DisclosureInfo:
      type: object
      properties:
        disclosed:
          type: integer
          format: int32
        total:
          type: integer
          format: int32
        complete:
          type: boolean
    VerificationResult:
      type: object
      properties:
        valid:
          type: boolean
        status:
          type: string
          enum:
          - VALID
          - TAMPERED
          - REVOKED
          - EXPIRED
          - NOT_FOUND
          - UNAVAILABLE
        credentialNumber:
          type: string
        reason:
          type: string
        issuerId:
          type: string
          format: uuid
        issuerName:
          type: string
        issuerDomain:
          type: string
        issuerVerified:
          type: boolean
        claims:
          type: object
          additionalProperties: {}
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
        anchorTxHash:
          type: string
        anchorBlockNumber:
          type: integer
          format: int64
        anchorChainId:
          type: integer
          format: int64
        anchorVerified:
          type: boolean
        disclosure:
          $ref: "#/components/schemas/DisclosureInfo"
    BatchItemResult:
      type: object
      properties:
        fileName:
          type: string
        credentialNumber:
          type: string
        recipientName:
          type: string
        issuerName:
          type: string
        status:
          type: string
        valid:
          type: boolean
        checks:
          $ref: "#/components/schemas/VerificationChecks"
        verificationResult:
          $ref: "#/components/schemas/VerificationResult"
        errorMessage:
          type: string
    BatchVerificationResponse:
      type: object
      properties:
        batchId:
          type: string
        total:
          type: integer
          format: int32
        processed:
          type: integer
          format: int32
        valid:
          type: integer
          format: int32
        tampered:
          type: integer
          format: int32
        revoked:
          type: integer
          format: int32
        expired:
          type: integer
          format: int32
        notFound:
          type: integer
          format: int32
        unavailable:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        durationMs:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchItemResult"
    VerificationChecks:
      type: object
      properties:
        envelopeStructure:
          type: boolean
        schemaConformance:
          type: boolean
        issuerSignature:
          type: boolean
        onChainAnchor:
          type: boolean
        revocationStatus:
          type: boolean
        ipfsIntegrity:
          type: boolean
...
    RecentAnchorResponse:
      type: object
      properties:
        credentialNumber:
          type: string
        issuerName:
          type: string
        blockNumber:
          type: integer
          format: int64
        anchoredAt:
          type: string
          format: date-time
    AnchorLookupResponse:
      type: object
      properties:
        credentialNumber:
          type: string
        contentHash:
          type: string
        txHash:
          type: string
        blockNumber:
          type: integer
          format: int64
        chainId:
          type: integer
          format: int64
        anchorVerified:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT